requirements.txt
groq>=0.9.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
